Move Strategy definition and implementation into new package
