+++ src/flower/app.py
@@ -13,6 +13,8 @@
 # limitations under the License.
 # ==============================================================================
 """Flower App."""
+
+
 from logging import DEBUG
 from typing import Dict
