+++ src/flower/strategy/default.py
@@ -1,0 +1,42 @@
+# Copyright 2020 Adap GmbH. All Rights Reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+# ==============================================================================
+"""Default strategy implementation."""
+
+
+from typing import Optional, Tuple
+
+from flower.typing import Weights
+
+from .strategy import Strategy
+
+
+class DefaultStrategy(Strategy):
+    """Strategy implementation used when no custom strategy is provided."""
+
+    def should_evaluate(self) -> bool:
+        """Evaluate every round."""
+        return True
+
+    def num_fit_clients(self, num_available_clients: int) -> Tuple[int, int]:
+        """Use 10% of available clients for training (minimum: 1)."""
+        return int(max(num_available_clients * 0.1, 1)), 1
+
+    def num_evaluation_clients(self, num_available_clients: int) -> Tuple[int, int]:
+        """Use 5% of available clients for evaluation (minimum: 1)."""
+        return int(max(num_available_clients * 0.05, 1)), 1
+
+    def evaluate(self, weights: Weights) -> Optional[Tuple[float, float]]:
+        """Do not evaluate."""
+        return None
